+++ src/components/features/cart/cart-check-book.tsx
@@ -1,8 +1,13 @@
 'use-client';
 import { Checkbox } from '@/components/ui/checkbox';
 import useCartStore from '@/store/cart/cart-store';
+import { TypeCartBooks } from '@/types/cart.type';
 
-export default function CartCheckBook({ id }: { id: string }) {
+type TypeCartCheckBookProps = {
+  id: TypeCartBooks['id'];
+};
+
+export default function CartCheckBook({ id }: TypeCartCheckBookProps) {
   const { checkedBooks, setCheckedBooks } = useCartStore();
   return (
     <Checkbox
